% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInterpretation.R
\name{getInterpretation}
\alias{getInterpretation}
\title{getInterpretation}
\usage{
getInterpretation()
}
\value{
interpretation
}
\description{
Get a brief, general interpretation of the descriptive statistics
generated by the makeDescriptives function.
}
\examples{
getInterpretation()


}
