% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumericDescriptives.R
\name{getNumericDescriptives}
\alias{getNumericDescriptives}
\title{getNumericDescriptives}
\usage{
getNumericDescriptives(y, name)
}
\arguments{
\item{y}{input structure}

\item{name}{name of structure to ID it in future}
}
\value{
descriptives, a vector of outputs
}
\description{
A function which pulls descriptive statistics for a single numeric input column
can be double or integer. Date types switched out to numeric, caution should be applied.
}
\examples{

stats <- getNumericDescriptives(myData, myName, myRounding)


}
