% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSuitable.R
\name{isSuitable}
\alias{isSuitable}
\title{isSuitable}
\usage{
isSuitable(y)
}
\arguments{
\item{y}{a data structure}
}
\value{
suitable, a boolean- TRUE if the structure is acceptable. Message: error message.
}
\description{
isSuitable
}
\examples{

isThisSuitable <- isSuitable(x)

}
