% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCharts.R
\name{makeCharts}
\alias{makeCharts}
\title{makeCharts}
\usage{
makeCharts(y, pName = NULL)
}
\arguments{
\item{pName.}{Optional, defaults to NULL, the path to put the charts in.}
}
\description{
A function that makes a wad of charts and puts them in your designated
working directory. The charts are histograms.
}
