% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDescriptives.R
\name{makeDescriptives}
\alias{makeDescriptives}
\title{Make Descriptives}
\usage{
makeDescriptives(y, roundDP = 2, fName = NULL)
}
\arguments{
\item{y}{the data to create summary statistics on}

\item{roundDP}{the number of decimal places to round relevant stats to}

\item{fName, }{optional parameter directing where to write out an xlsx file
with descriptive statistics. Defaults to NULL and file will be written to current
working directory}
}
\value{
tidy data frame with descriptive statistics
Also writes out a .xlsx file to designated location with the dataframe of
descriptive statistics.

If no filename or location specified, it will write to a file called
 "descriptives date/time.xlsx" in current working directory.
}
\description{
Make descriptive statistics in the style of Excel from your data frame.
}
\examples{
descriptives <- makeDescriptives(y)


}
