% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchStructure.R
\name{switchStructure}
\alias{switchStructure}
\title{switch Structure}
\usage{
switchStructure(y)
}
\arguments{
\item{y}{the column input. Can be vector or dataframe.}
}
\value{
x the column input with structure switched.
}
\description{
this function takes a single column input. If it is text-based, switches to factors
If it's a lubridate::date function, it will switch it to numeric.
}
